requirements.txt
# Core packages
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
PyYAML>=6.0

# Date and time handling
python-dateutil>=2.8.2

# Natural language processing (for sentiment analysis, etc.)
nltk>=3.6.0
regex>=2021.8.0

# Statistical analysis
scipy>=1.7.0
statsmodels>=0.13.0

# Visualization enhancements
plotly>=5.3.0
wordcloud>=1.8.0

# Progress tracking and CLI improvements
tqdm>=4.62.0
colorama>=0.4.4

# Optional: for advanced NLP capabilities
# transformers>=4.12.0
# torch>=1.10.0 